+++ packages/plugin-text/src/factory/types.ts
@@ -9,4 +9,5 @@ export interface TextPluginOptions {
 
 export type SlatePluginClosure = React.RefObject<{
   overlayContext: OverlayContextValue
+  name: string
 }>